Bootloader version and dependency updates
Update the TinyUF2 code to 0.10.2-43-ge91e98c
(Given what is actually used, this is practically no changes of any
relevance.)

Update the TinyUSB code to 0.14.0

Bump the version to v1.0

Any fixes made between now and product launch may change the build date,
but won't change the version number. Changes after product launch will
change the version number.

As part of this update, some experiments with FAVICON support were
conducted. It adds approximately 8KB to the build size, for which there
is room, but doesn't seem to add much benefit. It only works on Windows,
might make the icon more confusing, and it was ultimately decided to not
bother for now.