Implement measurement light calibration as part of gain calibration
Initially the whole measurement routine was written with the assumption
that a high gain (200ms) measurement could never saturate. This
assumption held true, with a comfortable margin, for all of the
prototype and pre-production units.

However, when calibration the first several production units, it was
discovered that the transmission light was bright enough to saturate a
high gain sensor measurement. To fix this problem, two approaches were
possible:
1) Redesign the target measurement routine to select from 3 gain
   settings instead of two.
2) Reduce the measurement light brightness if necessary

Option 1 sounds ideal, but implementing it while maintaining our
sensor-vs-LED synchronization requirements would have increased the
overall measurement time by at least 200ms.

Option 2 was therefore chosen, due to its minimal impact on the overall
target measurement process. Testing also showed that, due to the
logarithmic nature of the density scale, the extra brightness would not
have been that beneficial on high density targets.

The implementation involved adding an extra calibration parameter,
covering measurement brightness values for both reflection and
transmission. This was done for completeness, as it is not expected to
ever use a value less than 100% for reflection brightness.

The actual calibration process for this was folded into the gain
calibration routine, since the two steps are related and one is
necessary for the other to succeed.