Update default calibration values to better match expected targets
When selecting density values for calibrating a fresh device, the
selector menu starts with a default from which the user can select the
correct number. This change makes the defaults for some cases a lot
closer to the numbers expected on reference materials, thus making the
process a bit easier.