Change PWM frequency used for controlling LED brightness
This adjusts the timer prescaler value to change the PWM frequency from
125kHz down to 41.667kHz.

The reason for this change is that with the old frequency, at
sufficiently high duty cycle values, the low portion of the pulse did
not always drop below the LED driver's 0.5V threshold and did not stay
there for at least 100ns.

The result of violating these requirements, was that the brightness of
the connected LED would alternate between multiple values each time it
was turned on. This made bright (but not 100%) readings highly
inconsistent. This did not previously cause problems, but with the
addition of measurement brightness calibration, it does cause problems
now.