Switch all single-line TVS diodes to the Semtech uClamp3301H
The ST diodes used in the original design had two main problems:
- Their package size made handling and manufacturability difficult,
  resulting in a lot of spotted issues on production boards
- Their rating can only protect 5V-tolerant lines, and a few of the
  buttons are actually connected on MCU pins that may only be 3.3V
  tolerant

This changes all of these diodes to a Semtech diode that's in a
larger and easier to handle package, which is also designed for
protecting 3.3V lines. Care was taken to minimize any and all changes
not strictly necessary for changing this one part, in an effort to
preserve the validity of existing EMC certification. That being said,
the only arcing ever seen during that testing was to the shield of the
USB connector, which is handled by completely unrelated protection
components.

If it is later determined that a TVS diode with a 5V working voltage is
more appropriate, Semtech also makes the uClamp0501H which is in an
identical package and can be a drop-in alternative.

Fix #41