Update footprint origin of FFC connector J3
This moves the footprint origin by 1.975mm, so that it is over the body
of the component and not in the middle of the pins. This change was made
to roughly resemble other FFC connectors in the KiCad library, and might
result in a little less PCBA origin fiddling for assembly.

Ideally this just shifts some numbers around without making any actual
layout changes at all.

Fix #43