Update paste layer settings to optimize on a per-component level
The board layouts were initially using a -5% paste aperture reduction
globally. While this seemed to work okay for hand assembly, in
production it resulted in a few components not having enough paste on
their pads for a good solder connection.

Going forward, the plan is to now cover all pads 100% with paste unless
explicitly specified otherwise. This means setting the board-level paste
clearance to 0%, and making numerous component footprint updates.

Going forward, paste clearance shall only be adjusted on a per-component
basis. If that component's footprint is in the project library, then the
library footprint shall be updated. If that component's footprint is
from the KiCad library itself, then the adjustments may be made in the
PCB layout file.

This update is a starting point, making the most obvious updates. More
may be needed in the future prior to another production run.

Fix #42