Switch TVS diode from uClamp3301H to uClamp0501H
The 3301H has this annoying little feature called a snap-back voltage.
This means that if the voltage on the line ever exceeds the conservative
threshold of the diode, then it must fall below 2.8V before the diode
will stop conducting.

Given that transients associated with pushing a button will momentarily
exceed that threshold, this effectively results in a short across the
device's 3.3V power rail whenever a button is being held down. Thus, it
is safer to only protect buttons with TVS diodes that don't have this
feature.

Fix #41