Add ability to display density in stops
This mode will be called "f-stop" display for now, as it basically takes
the calculated density and converts it to a log2 format. To make it
clear what is being displayed, the number on the screen is prefixed
with a small "f/" indicator.

In this mode, 0.3D is roughly equivalent to f/1.0

It is still TBD whether this should affect anything beyond the local
device display for now.

Refs #50