USB Key Output behavior and config should track new display formats
The characters sent via the USB-HID feature now track the display format
settings, for both decimal separator and measurement unit.

To avoid character conflicts, a field separator setting of ',' is now
automatically changed to a ';' if a comma decimal separator is selected.
No new setting value has been added here, so the same field separator
selection is implemented as either character depending on the decimal
separator.

The USB-CDC protocol is as of yet unchanged, since it is assumed that
this is to be interpreted entirely by software which is capable of
implementing any format changes itself. It is also desired to avoid
breaking any existing software.